scrappages/requirements.txt
django==4.1.7
djangorestframework==3.14.0
gunicorn==20.1.0
python-decouple==3.8
psycopg2-binary==2.9.5
Pillow
